+++ pybolano/core/__init__.py
@@ -1,8 +1,5 @@
 from pybolano.core.commutators import (
-    NO_commutator,
-    expand_comm_A_BC,
-    expand_comm_AB_C,
-    expand_comm_AB_CD,
+    NO_commutator
 )
 from pybolano.core.Lindblad_ME import (
     Hamiltonian_trace,
@@ -13,9 +10,6 @@ from pybolano.core.normal_ordering import NO, normal_ordering
 
 __all__ = [
     "NO_commutator",
-    "expand_comm_A_BC",
-    "expand_comm_AB_C",
-    "expand_comm_AB_CD",
     "normal_ordering",
     "Hamiltonian_trace",
     "LME_expval_evo",